Added javadoc comments to DamageGUIController class and edited some comments in both DamageGUIController and CharacterGUIController classes. removed some unnecessary whitespaces. Added some TODOs for more features that still need to be implemented later on. Renamed damageGUI.java and damageGUI.fxml to uppercase first letter to better match the Java conventions. Added a missing 'break;' statement to Main.chooseScene() method.
